refactor(channels): move discord_http -> channels/discord/http
Slice 3 of the phase-1a lane. Shared 429/backoff HTTP helper is Discord
channel mechanics. Pure move + dotted-import rewrite in 5 importers + 2
test loaders. One line touches #3239-contested discord_rest_client.py
(import only), resolved at the post-merge rebase.
